+++ src/Elders.Cronus.Tests/InMemoryEventStoreSuite/When_loading_aggregate_root_with_entity_from_event_store.cs
@@ -1,3 +1,5 @@
+using Elders.Cronus.AtomicAction;
+using Elders.Cronus.AtomicAction.InMemory;
 using Elders.Cronus.DomainModeling;
 using Elders.Cronus.EventStore;
 using Elders.Cronus.EventStore.InMemory;