+++ src/Elders.Cronus.Tests/InMemoryEventStoreSuite/When_saving_aggregate_root_in_event_store.cs
@@ -3,6 +3,8 @@ using Elders.Cronus.Tests.TestModel;
 using Machine.Specifications;
 using Elders.Cronus.EventStore.InMemory;
 using Elders.Cronus.EventStore;
+using Elders.Cronus.AtomicAction;
+using Elders.Cronus.AtomicAction.InMemory;
 
 namespace Elders.Cronus.Tests.InMemoryEventStoreSuite
 {