+++ src/Elders.Cronus.Tests/MessageStreaming/When__MessageStream__is_feeded_with_failed_messages_with_origin_type__handler__.cs
@@ -39,7 +39,6 @@ namespace Elders.Cronus.Tests.MessageStreaming
 
         It should_report_about_all_successes = () => secondFeedResult.SuccessfulMessages.Count().ShouldEqual(numberOfMessages);
 
-        static IFeedResult firstFeedResult;
         static IFeedResult secondFeedResult;
         static int numberOfMessages = 3;
         static MessageProcessor messageStream;