+++ src/Elders.Cronus.Tests/TestModel/TestAggregateRoot.cs
@@ -1,4 +1,3 @@
-using System;
 using Elders.Cronus.DomainModeling;
 
 namespace Elders.Cronus.Tests.TestModel