+++ src/Elders.Cronus/DomainModeling/IAggregateVersionService.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Elders.Cronus.DomainModeling
-{
-    public interface IAggregateRevisionService : IDisposable
-    {
-        int ReserveRevision(IAggregateRootId aggregateId, int requestedRevision);
-    }
-}