+++ src/Elders.Cronus/Pipeline/CircuitBreaker/IEndpointConsumerErrorStrategy.cs
@@ -1,5 +1,3 @@
-using Elders.Cronus.DomainModeling;
-
 namespace Elders.Cronus.Pipeline.CircuitBreaker
 {
     public interface ICircuitBreakerErrorStrategy