+++ src/Elders.Cronus/Pipeline/Config/PipelineConsumerSettings.cs
@@ -6,6 +6,7 @@ using Elders.Cronus.Pipeline.Hosts;
 using Elders.Cronus.Pipeline.Transport;
 using Elders.Cronus.Serializer;
 using Elders.Cronus.EventStore;
+using Elders.Cronus.AtomicAction;
 
 namespace Elders.Cronus.Pipeline.Config
 {
@@ -48,9 +49,9 @@ namespace Elders.Cronus.Pipeline.Config
             base.Build();
 
             var builder = this as ISettingsBuilder;
-            builder.Container.RegisterSingleton<IAggregateRootAtomicAction>(() => new InMemoryAggregateRootAtomicAction(), builder.Name);
-            var aggregateRepository = new AggregateRepository(builder.Container.Resolve<IEventStore>(builder.Name), builder.Container.Resolve<IAggregateRootAtomicAction>(builder.Name));
-            builder.Container.RegisterSingleton<IAggregateRepository>(() => aggregateRepository, builder.Name);
+            Func<IAggregateRootAtomicAction> atomicAction = () => builder.Container.Resolve<IAggregateRootAtomicAction>();
+            Func<IAggregateRepository> aggregateRepository = () => new AggregateRepository(builder.Container.Resolve<IEventStore>(builder.Name), atomicAction());
+            builder.Container.RegisterSingleton<IAggregateRepository>(() => aggregateRepository(), builder.Name);
         }
     }