+++ src/Elders.Cronus/Publisher.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using Elders.Cronus.DomainModeling;
+using Elders.Cronus.Logging;
 
 namespace Elders.Cronus
 {
     public abstract class Publisher<TMessage> : IPublisher<TMessage> where TMessage : IMessage
     {
-        static readonly log4net.ILog log = log4net.LogManager.GetLogger(typeof(Publisher<TMessage>));
+        static readonly ILog log = LogProvider.GetLogger(typeof(Publisher<TMessage>));
 
         protected abstract bool PublishInternal(TMessage message, Dictionary<string, string> messageHeaders);
 
@@ -15,13 +16,12 @@ namespace Elders.Cronus
             try
             {
                 PublishInternal(message, messageHeaders);
-                if (log.IsInfoEnabled)
-                    log.Info("PUBLISH => " + message);
+                log.Info(() => "PUBLISH => " + message);
                 return true;
             }
             catch (Exception ex)
             {
-                log.Error(ex.Message, ex);
+                log.ErrorException(ex.Message, ex);
                 return false;
             }
         }